+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Updated OpenTelemetry modules to [v1.42.0/v0.136.0](https://github.com/open-telemetry/opentelemetry-collector/releases/tag/v0.136.0)
 - Add `mqttreceiver`
 - `swohostmetricsreceiver` utilizes OTEL generated code instead of original internal implementation. Relevant code cleanup.
+- Added receiver `swok8sdiscovery` to publish known database entities and relationships
 
 ## v0.131.9
 - Updates golang to 1.25.1